Fix moving files into new dirs
Matt Krai reported a problem in Debian bug #293088 with moving an obsoleted
file into a new location where the parent of the new file did not exist.
Darcs was being given a darcs add command to add the parent directory,
but did not like being given an absolute path.  Code adjusted to chdir,
then issue a relative darcs add.  Testing showed this fixed the problem for
me.