Now support summary in git, debian #403505
From: Guido Guenther
Hi,
currently the summary message can be given on the command line but it's
silently being ignored. Since the summary in git is the first line of
the commit message we simply prepend the summary if it's given on the
command line otherwise we take what's autogenerated.
Please apply,
 -- Guido